Alteração após SonarLint

+++ tarefa009/Exerc02/Program.cs
@@ -14,7 +14,7 @@ namespace Exerc2
         {
             Console.WriteLine("A média de 8, 9 e 7 é: " + calculaMedia(8, 9, 7));
             Console.WriteLine("A média de 4, 5 e 6 é: " + calculaMedia(4, 5, 6));
-            Console.WriteLine("A soma dessas duas médias é: " + (int)(calculaMedia(4, 5, 6) + calculaMedia(8, 9, 7)));
+            Console.WriteLine("A soma dessas duas médias é: " + (calculaMedia(4, 5, 6) + calculaMedia(8, 9, 7)));
             Console.WriteLine("A média dessas duas médias é: " + (double)(calculaMedia(4, 5, 6) + calculaMedia(8, 9, 7)) / 2);
         }
     }
